pull out autoconfig into separate function
